1. Where are you? mandatory How do you get the variable identifier (which is the memory address in the CPython implementation)?
Write the name of the function in the file, without ().

0. Who am I? mandatory
What function would you use to print the type of an object?

Write the name of the function in the file, without ().

2. Right count mandatory
In the following code, do a and b point to the same object? Answer with Yes or No.

>>> a = 89
>>> b = 100

3. Right count = mandatory
In the following code, do a and b point to the same object? Answer with Yes or No.

>>> a = 89
>>> b = 89

4. Right count = mandatory
In the following code, do a and b point to the same object? Answer with Yes or No.

>>> a = 89
>>> b = a

5. Right count =+ mandatory
In the following code, do a and b point to the same object? Answer with Yes or No.
>>> a = 89
>>> b = a + 1

6. Is equal mandatory
What do these 3 lines print?

>>> s1 = Holberton
>>> s2 = s1
>>> print(s1 == s2)