20. Tuple or not? mandatory a = () Is a a tuple? Answer with Yes or No.
21. Tuple or not? mandatory
a = (1, 2)
Is a a tuple? Answer with Yes or No.

22. Tuple or not? mandatory
a = (1)
Is a a tuple? Answer with Yes or No.

23. Tuple or not? mandatory
a = (1, )
Is a a tuple? Answer with Yes or No.

24. Richard Sim's special #0 mandatory
What does this script print?

a = (1)
b = (1)
a is b

25. Richard Sim's special #1 mandatory
What does this script print?

a = (1, 2)
b = (1, 2)
a is b

What does this script print?

a = ()
b = ()
a is b
Repo:

GitHub repository: holbertonschool-higher_level_programming
Directory: 0x09-python-everything_is_object
File: 26-answer.txt

27. Richard Sim's special #3 mandatory
>>> id(a)
139926795932424
>>> a
[1, 2, 3, 4]
>>> a = a + [5]
>>> id(a)
Will the last line of this script print 139926795932424? Answer with Yes or No.

28. Richard Sim's special #4 mandatory
>>> a
[1, 2, 3]
>>> id (a)
139926795932424
>>> a += [4]
>>> id(a)
Will the last line of this script print 139926795932424? Answer with Yes or No.

30. Clear strings mandatory
guillaume@ubuntu:/python3$ cat string.py
a = HBTN
b = HBTN
del a
del b
c = HBTN
guillaume@ubuntu:/python3$
Assuming we are using a CPython implementation of Python3 with default options/configuration (For answers with numbers use integers, don’t spell out the word):

How many string objects are created by the execution of the first line of the script? (106-line1.txt)
How many string objects are created by the execution of the second line of the script (106-line2.txt)
After the execution of line 3, is the string object pointed by a deleted? Answer with Yes or No (106-line3.txt)
After the execution of line 4, is the string object pointed by b deleted? Answer with Yes or No (106-line4.txt)
How many string objects are created by the execution of the last line of the script (106-line5.txt)